Uploaded the initial Dicebold method.
Uploaded a simple Python script that contains the primary functionality of the bot. Eventually this method will be cordoned off and called by the program whenever a user requests it. This method contains the core functionality that will serve 90% of the bot's uses; roll a set number of a set type of die, and add or subtract an inputted integer from it.